refactor: better naming to command storage abstraction file name
