+++ commands/build.command.ts
@@ -2,7 +2,7 @@ import { Command, CommanderStatic } from 'commander';
 import type { BuildAction } from '../actions';
 import { ERROR_PREFIX, INFO_PREFIX } from '../lib/ui';
 import { AbstractCommand } from './abstract.command';
-import { CommandStorage } from './command.input';
+import { CommandStorage } from './command-storage';
 
 export class BuildCommand extends AbstractCommand<BuildAction> {
   public load(program: CommanderStatic): void {
