+++ commands/index.ts
@@ -1,2 +1,2 @@
 export * from './command.loader';
-export * from './command.input';
+export * from './command-storage';
